refactor: hoist protocol + crypto + signing to src/shared
Move src/relay/protocol.ts, src/broker/crypto.ts, and src/broker/signing.ts
into src/shared/ so the upcoming src/client/ library can import them without
crossing relay/broker subdirectory boundaries. All import sites updated.
