+++ src/shared/crypto.ts
