+++ src/shared/signing.ts
